feat: Added functions to visualize the rectangles around the entities
Before we add the code to detect collissions among our entities, we may want
to visualize the rectangles/bounds around them first. Using the image width
and height is not feasible in our case because there are lots of padding in our
images. It means that the width and height of our images are not the same as
the actual width and height of the visible entity.

With this in mind we added a "bounds" property for our entity that will tell
us the position of our entity inside the image. It has x, y, width and height
that is relative to its image. We then draw it as a rectangle to help visualize
them on the screen.

After perfecting the bounds of each entities, we can now use these bound objects
to detect collissions between the player and the enemies.