feat: Added collision detection and allowed multiple key movement
When the bounds of the player or any of the ememy overlaps, the location of the
player is reset to its initial y location and random x location.

The movement is now not one big leap anymore. The player has been assigned a
predefined speed and it can move in any one or two directions at a time. This
has been made possible by monitoring which movement keys are pressed at any one
time and updating its position based on it. So now, when the user press both
up and left buttons at the same time, the player will move to the north west
direction.