refactor: Made the context, and grid properties available globally
The canvas context can now be accessed by calling Resources.getContext() and
set by calling Resources.setContext(ctx).

The grid block rows are columns are also moved to Resources.js and the block
types are converted from whole words to just the capital first letter of the
block. That is: G is for grass, S is for stone and W is for water. We can now
configure the rows in a string such as "WSSSGG". It means the first row is
Water, the 2nd to 4th rows are Stone, and the last 2 rows are Grass.

A new functionn is created to select a random row when the Entity is reset.
When a player is reset it will select a random Grass row block for its Y pos.
When an enemy is reset it will select a random Stone row block for its Y pos.

All entity functions are now updated to remove the context, rows, and columns
parameter since they are now globally accessible by calling Resources.getGrid()
and Resources.getContext().