feat: Added Game Over check and display remaining lives
The player will have 3 lives as indicated by the Hearts in the upper right hand
side of the screen. A heart will fade if a life was lost. When all lives are lost
the game is over and the user will have to press the ESC key to play a new game.

Collision checks will not be made when the game is over as the Game Over message
displays at the center of the screen.