refactor: Made the Heart inherit from Entity; Changed goal animation.
The Heart entity now inherits from Entity so that it will not need to
implement its own render method.

The goal animation is changed from changing the opacity to bouncing/
jumping animation.

I also made the row blocks to default since we are not yet implementing
the game levels which will change the row blocks in each level and add
more entities to the game.