feat: Added collectibles and levels
Added gem, star and key collectibles. The gem will give you extra 1 point.
The star will give you 2 points and 1 extra life. The key will give you
3 points and will unlock the new block row arrangements (The water, stone,
and grass rows will be shuffled).

Depending on the level, the collectibles will be added on the screen. The
current level is now displayed at the top center on the screen.