fix: Enemy and water collision bug fix
Before, when the player collide with 2 enemies, 2 lives are lost.
Now this is fixed.

Another bug is when the water block is in the bottom of the screen,
even the player is still halfway to its block, it already align itself
to that block. Now, the player should be close enough to the water row
before it will be aligned properly.